+++ src/routes/forms/forms.tsx
@@ -7,52 +7,12 @@ import { CountryEnum } from "../../enums/country.enum";
 import { GenderEnum } from "../../enums/gender.enum";
 import { users } from "../../db/dbusers";
 import Card from "../home/Card/card";
-import {initialState} from "./initialState";
-
-export interface FormControlRadio {
-  ref: React.RefObject<HTMLInputElement>[];
-  type: string;
-  label: GenderEnum[];
-  validation?: ValidationControl;
-  errorMessage?: string;
-  valid: boolean;
-}
-
-export interface FormControlSelect {
-  ref: React.RefObject<HTMLSelectElement>;
-  type: string;
-  label: string;
-  options: CountryEnum[];
-  validation?: ValidationControl;
-  errorMessage?: string;
-  valid: boolean;
-}
-
-export interface FormControlInput {
-  ref: React.RefObject<HTMLInputElement>;
-  type: string;
-  label: string;
-  validation?: ValidationControl;
-  errorMessage?: string;
-  valid: boolean;
-}
-
-export interface ValidationControl {
-  required: boolean;
-  firstLetter?: boolean;
-}
-
-export type ItemFormType = 'name' | 'surname' | 'birthday' | 'country' | 'gender' | 'married' | 'photo';
-
-export interface IForm {
-  name: FormControlInput;
-  surname: FormControlInput;
-  birthday: FormControlInput;
-  country: FormControlSelect;
-  married: FormControlInput;
-  gender: FormControlRadio;
-  photo: FormControlInput;
-}
+import { initialState } from "./initialState";
+import {
+  IForm,
+  ItemFormType,
+  ValidationControl,
+} from "interfaces/form.interface";
 
 function validate(
   validation: ValidationControl,
@@ -72,14 +32,17 @@ function validate(
   return true;
 }
 
-class UserForm extends React.Component<unknown, { formState: IForm }> {
-
+class UserForm extends React.Component<
+  unknown,
+  { formState: IForm; isHasPopup: boolean }
+> {
   public formRef: React.RefObject<HTMLFormElement>;
   constructor(props: unknown) {
     super(props);
     this.formRef = React.createRef();
     this.state = {
-      formState: {...initialState},
+      formState: { ...initialState },
+      isHasPopup: false,
     };
     this.handleSubmit = this.handleSubmit.bind(this);
   }
@@ -101,30 +64,30 @@ class UserForm extends React.Component<unknown, { formState: IForm }> {
       const keyValue: ItemFormType = value[0] as ItemFormType;
       if (!!value[1]?.validation) {
         updatedState[keyValue].valid = validate(
-            value[1]?.validation,
-            value[1].ref?.current?.value
+          value[1]?.validation,
+          value[1].ref?.current?.value
         );
       }
       switch (keyValue) {
         case "name":
-            user[keyValue] = value[1].ref?.current?.value;
+          user[keyValue] = value[1].ref?.current?.value;
           break;
         case "surname":
-            user[keyValue] = value[1].ref?.current?.value;
+          user[keyValue] = value[1].ref?.current?.value;
           break;
         case "birthday":
-            user[keyValue] = value[1].ref?.current?.value;
+          user[keyValue] = value[1].ref?.current?.value;
           break;
         case "country":
-            user[keyValue] = value[1].ref?.current?.value;
+          user[keyValue] = value[1].ref?.current?.value;
           break;
         case "married":
-            user[keyValue] = value[1].ref?.current?.checked;
+          user[keyValue] = value[1].ref?.current?.checked;
           break;
         case "photo":
-            user[keyValue] = value[1].ref?.current?.files[0]
-              ? URL.createObjectURL(value[1].ref?.current?.files[0])
-              : "";
+          user[keyValue] = value[1].ref?.current?.files[0]
+            ? URL.createObjectURL(value[1].ref?.current?.files[0])
+            : "";
           break;
       }
     });
@@ -137,19 +100,18 @@ class UserForm extends React.Component<unknown, { formState: IForm }> {
       }
       return ref.current?.checked;
     });
-    if (user.gender === GenderEnum.XZ) {
-      updatedState.gender.valid = false;
-    } else {
-      updatedState.gender.valid = true;
-    }
+    updatedState.gender.valid = user.gender !== GenderEnum.XZ;
     const isInvalidForm = Object.values(updatedState).some(
       (control) => !control.valid
     );
     if (!isInvalidForm) {
       user.id = `${Math.random() * 10000}`;
       users.push(user);
-      this.setState({ formState: {...initialState} });
-      this.formRef?.current?.reset()
+      this.setState({ formState: { ...initialState }, isHasPopup: true });
+      setTimeout(() => {
+        this.setState({ isHasPopup: false });
+      }, 3000);
+      this.formRef?.current?.reset();
     } else {
       this.setState({ formState: updatedState });
     }
@@ -158,84 +120,92 @@ class UserForm extends React.Component<unknown, { formState: IForm }> {
   render() {
     return (
       <div>
-        <div         style={{
-          display: "flex",
-          padding: "20px",
-          justifyContent: "space-between",
-          flexWrap: "wrap",
-          border: "1px solid red",
-        }}>
+        <div
+          style={{
+            padding: "20px",
+            border: "1px solid red",
+          }}
+        >
           <form ref={this.formRef} onSubmit={this.handleSubmit}>
             <div
-                style={{
-                  margin: "30px",
-                }}
+              style={{
+                margin: "30px",
+              }}
             >
               <CustomInput formControl={this.state.formState.name} />
             </div>
 
             <div
-                style={{
-                  margin: "30px",
-                }}
+              style={{
+                margin: "30px",
+              }}
             >
               <CustomInput formControl={this.state.formState.surname} />
             </div>
 
             <div
-                style={{
-                  margin: "30px",
-                }}
+              style={{
+                margin: "30px",
+              }}
             >
               <CustomInput formControl={this.state.formState.birthday} />
             </div>
 
             <div
-                style={{
-                  margin: "30px",
-                }}
+              style={{
+                margin: "30px",
+              }}
             >
               <CustomSelect formControl={this.state.formState.country} />
             </div>
 
             <div
-                style={{
-                  margin: "30px",
-                }}
+              style={{
+                margin: "30px",
+              }}
             >
               <CustomRadio formControl={this.state.formState.gender} />
             </div>
 
             <div
-                style={{
-                  margin: "30px",
-                }}
+              style={{
+                margin: "30px",
+              }}
             >
               <CustomInput formControl={this.state.formState.married} />
             </div>
 
             <div
-                style={{
-                  margin: "30px",
-                }}
+              style={{
+                margin: "30px",
+              }}
             >
               <CustomInput formControl={this.state.formState.photo} />
             </div>
 
             <button type="submit">Submit</button>
           </form>
+          {this.state.isHasPopup ? (
+            <h2
+              style={{
+                border: "2px solid green",
+              }}
+            >
+              Card was created
+            </h2>
+          ) : null}
         </div>
         <div
-            style={{
-              display: "flex",
-              padding: "20px",
-              gap: "10px",
-              flexWrap: "wrap",
-              border: "1px solid red",
-            }}
+          style={{
+            display: "flex",
+            padding: "20px",
+            gap: "10px",
+            flexWrap: "wrap",
+            border: "1px solid red",
+          }}
         >
           {users.map((card: IUser) => (
-              <Card key={card.id} card={card} />
+            <Card key={card.id} card={card} />
           ))}
         </div>
       </div>
